+++ builtin/providers/heroku/resource_heroku_drain.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/bgentry/heroku-go"
 	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform/terraform"
 )
 
 func resourceHerokuDrain() *schema.Resource {
@@ -51,6 +52,9 @@ func resourceHerokuDrainCreate(d *schema.ResourceData, meta interface{}) error {
 	d.SetId(dr.Id)
 	d.Set("url", dr.URL)
 	d.Set("token", dr.Token)
+	d.SetDependencies([]terraform.ResourceDependency{
+		terraform.ResourceDependency{ID: app},
+	})
 
 	log.Printf("[INFO] Drain ID: %s", d.Id())
 	return nil